Move `SpirvSource` out of CLI
