+++ crates/cargo-gpu/Cargo.toml
@@ -29,7 +29,6 @@ dunce.workspace = true
 tempfile.workspace = true
 test-log.workspace = true
 cargo_metadata = { workspace = true, features = ["builder"] }
-cargo-util-schemas.workspace = true
 
 [lints]
 workspace = true
