+++ crates/cargo-gpu/src/target_specs.rs
@@ -26,9 +26,10 @@ use std::path::{Path, PathBuf};
 
 use anyhow::Context as _;
 use cargo_metadata::Metadata;
-use rustc_codegen_spirv_cache::cache::cache_dir;
-
-use crate::spirv_source::{FindPackage as _, SpirvSource};
+use rustc_codegen_spirv_cache::{
+    cache::cache_dir,
+    spirv_source::{FindPackage as _, SpirvSource},
+};
 
 /// Extract legacy target specs from our executable into some directory
 pub fn write_legacy_target_specs(target_spec_dir: &Path) -> anyhow::Result<()> {
