+++ crates/rustc_codegen_spirv-cache/Cargo.toml
@@ -9,7 +9,14 @@ license.workspace = true
 
 [dependencies]
 thiserror.workspace = true
+anyhow.workspace = true
+log.workspace = true
 directories.workspace = true
+cargo_metadata.workspace = true
+
+[dev-dependencies]
+test-log.workspace = true
+cargo-util-schemas.workspace = true
 
 [lints]
 workspace = true
